new recref syntax implemented
