+++ unit/CMakeLists.txt
@@ -1,9 +1,12 @@
 cmake_minimum_required(VERSION 2.8)
+include(${CMAKE_SOURCE_DIR}/xenonunit.txt)
+
+add_executable(recref recref_unit.cpp)
+add_test(recref recref)
+
 add_subdirectory(xddlunit)
 add_subdirectory(itu)
 add_subdirectory(xmlparse)
 add_subdirectory(document)
 add_subdirectory(script)
-
-enable_testing()
