+++ unit/document/specunit.h
@@ -29,7 +29,6 @@ class doc_unit
 
         ut.add(&doc_unit::fieldtypes);
         ut.add(&doc_unit::allTypes);
-        ut.add(&doc_unit::recref_regex);
         ut.add(&doc_unit::get_record);
         ut.skip();
         ut.cont();
@@ -53,4 +52,5 @@ class doc_unit
     void allTypes();
     void recref_regex();
     void get_record();
+    void relative_url();
 };
