+++ unit/itu/ituunit.cpp
@@ -132,72 +132,6 @@ void itu_unit::sanity()
     IT_ASSERT(c == b);
 }
 
-struct url_unit_type {
-    std::string str() const {
-        return path + file + anchor;
-    }
-
-    std::string path;
-    std::string file;
-    std::string anchor;
-};
-
-bool operator==(const xenon::recref & x, const url_unit_type & y) {
-    return (x.path == y.path) && (x.file == y.file) && (x.anchor == y.anchor);
-}
-
-void itu_unit::create_url()
-{
-    std::vector< std::pair<xenon::recref, url_unit_type> > url_tests = 
-        { 
-            { { "#4d" }, {"", "", "#4d"} },
-            { { "TS-23.038.xddl#4d" }, {"", "TS-23.038.xddl", "#4d"} },
-            { { "3GPP2/TS-23.038.xddl#4d" }, {"3GPP2/", "TS-23.038.xddl", "#4d"} },
-            { { "a/3GPP2/TS-23.038.xddl#4d" }, {"a/3GPP2/", "TS-23.038.xddl", "#4d"} },
-        };
-
-    for (const auto & i : url_tests) { // vs2014 won't compile this without this brace
-        IT_ASSERT_MSG(i.first.str() << " != " << i.second.str(), i.first == i.second);
-    }
-}
-
-struct relative_url_type {
-    xenon::recref base;
-    xenon::recref offset;
-    xenon::recref result;
-};
-    
-
-void itu_unit::relative_url() {
-    std::vector<relative_url_type> url_tests = {
-        { "", "" , "" },
-        { "", "#3.1.1" , "#3.1.1" },
-        { "", "C.R1001-G.xddl" , "C.R1001-G.xddl" },
-        { "", "C.R1001-G.xddl#3.1.1" , "C.R1001-G.xddl#3.1.1" },
-
-        { "C.S0005.xddl", "" , "C.S0005.xddl" },
-        { "C.S0005.xddl", "#3.1.1" , "C.S0005.xddl#3.1.1" },
-        { "C.S0005.xddl", "C.R1001-G.xddl" , "C.R1001-G.xddl" },
-        { "C.S0005.xddl", "C.R1001-G.xddl#3.1.1" , "C.R1001-G.xddl#3.1.1" },
-
-        { "3GPP2/C.S0005.xddl", "" , "3GPP2/C.S0005.xddl" },
-        { "3GPP2/C.S0005.xddl", "#3.1.1" , "3GPP2/C.S0005.xddl#3.1.1" },
-        { "3GPP2/C.S0005.xddl", "C.R1001-G.xddl" , "3GPP2/C.R1001-G.xddl" },
-        { "3GPP2/C.S0005.xddl", "C.R1001-G.xddl#3.1.1" , "3GPP2/C.R1001-G.xddl#3.1.1" },
-
-        { "3GPP2/C.S0005.xddl", "foo/b.xddl" , "3GPP2/foo/b.xddl" },
-
-        { "icd.xddl", "#zero", "icd.xddl#zero" },
-        { "3GPP/TS-24.008.xddl", "TS-24.007.xddl#11.2", "3GPP/TS-24.007.xddl#11.2" },
-        { "3GPP/TS-24.008.xddl", "TS-24.007.xddl#11.2.2.1c", "3GPP/TS-24.007.xddl#11.2.2.1c" },
-    };
-
-    for (const auto & i : url_tests) {
-        auto result = xenon::relative_url(i.base, i.offset);
-        IT_ASSERT_MSG(result.str() << " == " << i.result.str(), result == i.result);
-    }
-}
-
 int main (int, char **) {
     itu_unit test;
     ict::unit_test<itu_unit> ut(&test);
