+++ unit/itu/ituunit.h
@@ -10,8 +10,6 @@ class itu_unit {
         ut.add(&itu_unit::bit_lengths);
         ut.add(&itu_unit::splits);
         ut.add(&itu_unit::sanity);
-        ut.add(&itu_unit::create_url);
-        ut.add(&itu_unit::relative_url);
     }
 
     /* Tests */
@@ -19,7 +17,5 @@ class itu_unit {
     void bit_lengths();
     void splits();
     void sanity();
-    void create_url();
-    void relative_url();
 };
