+++ xenonunit.txt
@@ -3,3 +3,4 @@ include_directories(${CMAKE_SOURCE_DIR}/include)
 if(NOT WIN32)
 add_definitions(-std=c++11)
 endif()
+enable_testing()
